change to use findOperators
