+++ tests/src/test/scala/com/nvidia/spark/rapids/TestUtils.scala
@@ -61,27 +61,6 @@ object TestUtils extends Assertions with Arm {
         actual.column(i).asInstanceOf[GpuColumnVector].getBase)
     }
   }
-  
-  /** Return list of  matching predicates present in the plan */
-  def findOperators(plan: SparkPlan, predicate: SparkPlan => Boolean): Seq[SparkPlan] = {
-    def recurse(
-      plan: SparkPlan,
-      predicate: SparkPlan => Boolean,
-      accum: ListBuffer[SparkPlan]): Seq[SparkPlan] = {
-      plan match {
-        case _ if predicate(plan) =>
-          accum += plan
-          plan.children.flatMap(p => recurse(p, predicate, accum)).headOption
-        case a: AdaptiveSparkPlanExec => recurse(a.executedPlan, predicate, accum)
-        case qs: BroadcastQueryStageExec => recurse(qs.broadcast, predicate, accum)
-        case qs: ShuffleQueryStageExec => recurse(qs.shuffle, predicate, accum)
-        case other => other.children.flatMap(p => recurse(p, predicate, accum)).headOption
-      }
-      accum
-    }
-
-    recurse(plan, predicate, new ListBuffer[SparkPlan]())
-  }
 
   /** Return final executed plan */
   def getFinalPlan(plan: SparkPlan): SparkPlan = {
